RAG with LlamaIndex - Project 3/requirements.txt
llama-index
llama-index-embeddings-cohere
llama-index-llms-cohere
llama-index-vector-stores-chroma
chromadb
cohere
gradio
python-dotenv
